+++ drypy/utils.py
@@ -8,13 +8,32 @@
 
 import logging
 
+from drypy import get_logging_level
+
+
 logger = logging.getLogger(__name__)
 
 
 def log_call(func, *args, **kw):
     """Produce a info message which logs the *func* call.
 
+    Args:
+        func: callable whose call is to be logged
+        _drypy_log_level: custom logging level to override the global
+            one
+        _drypy_custom_msg: custom message to be logged instead of the
+            default one
+        *args: func's positional arguments
+        **kw: func's keyword arguments
+
     """
+    # pop any system arguments
+    log_level = kw.pop('_drypy_log_level', get_logging_level())
+    custom_msg = kw.pop('_drypy_custom_msg', None)
+    if custom_msg is not None:
+        logger.log(log_level, custom_msg)
+        return
+
     # concatenate args and kw args transforming string values
     # from 'value' to '"value"' in order to pretty display em
     func_args = []
@@ -38,4 +57,4 @@ def log_call(func, *args, **kw):
     # print the log message
     msg = "[DRYRUN] call to '{func}({args})'"
     msg = msg.format(func=func.__name__, args=", ".join(func_args))
-    logger.info(msg)
+    logger.log(log_level, msg)