+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [2025-04-18] 1.2.0
+
+### Added
+- now it's possible to set the global (#24)
+- now it's possible to pass args to `sham` decorator (#26)
+  - `log_level` custom log level just for the decorated function
+  - `custom_msg` custom log message to replace default output
+  - `return_value` to return a custom value instead of None, in order to
+    emulate and replace the `sentinel` decorator
+
+
 ## [2025-01-04] 1.1.0
 
 ### Added